requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
passlib~=1.7.4
pymongo==4.6.0
motor~=3.3.1
python-dotenv~=1.0.0
aiofiles~=23.2.1
asyncio~=3.4.3
jose~=1.0.0
requests~=2.31.0
aiohttp~=3.8.5
Pillow==10.1.0
Faker==19.13.0
pytest==7.4.3
httpx==0.25.1
reportlab~=4.0.5
fpdf~=1.7.2
gradio==4.1.1
